Moving thread_pool out of experimental
