+++ dhorn.vcxproj.filters
@@ -186,13 +186,13 @@
     <ClInclude Include="include\dhorn\console.h">
       <Filter>include</Filter>
     </ClInclude>
-    <ClInclude Include="include\dhorn\experimental\thread_pool.h">
-      <Filter>include\experimental</Filter>
-    </ClInclude>
     <ClInclude Include="include\dhorn\debug.h">
       <Filter>include</Filter>
     </ClInclude>
     <ClInclude Include="include\dhorn\functional.h">
+      <Filter>include</Filter>
+    </ClInclude>
+    <ClInclude Include="include\dhorn\thread_pool.h">
       <Filter>include</Filter>
     </ClInclude>
   </ItemGroup>
